Extract homedemo to separate component
